Updated {Oneshot,Broadcast}Channel::close signature
Oneshot & broadcast channels now return `CloseStatus` when
`close` is called, which allows the user to know if the
channel was already closed.
